모듈화
카테고리 별 JS문서에 공통적으로 해당되는 기능 JS폴더에 movie.js
폴더로 옮겨 모듈화 했습니다.
