+++ nowPlaying/now_playing.html
@@ -12,6 +12,7 @@
     />
     <title>GOE's CINEMA</title>
     <link rel="stylesheet" type="text/css" href="../css/html.css" />
+    <script src="/nowPlaying/nowPlaying.js" type="module"></script>
   </head>
 
   <body>
@@ -76,6 +77,5 @@
       </div>
     </footer>
     <div class="moveTopBtn"><img src="../img/top.png" alt="" /></div>
-    <script src="/nowPlaying/nowPlaying.js"></script>
   </body>
 </html>
