+++ popular/popular.html
@@ -12,6 +12,7 @@
     />
     <title>GOE's CINEMA</title>
     <link rel="stylesheet" type="text/css" href="../css/html.css" />
+    <script src="/popular/popular.js" type="module"></script>
   </head>
 
   <body>
@@ -78,6 +79,5 @@
       </div>
     </footer>
     <div class="moveTopBtn"><img src="../css/top.png" alt="" /></div>
-    <script src="/popular/popular.js"></script>
   </body>
 </html>
