+++ topRated/top_rated.html
@@ -12,6 +12,7 @@
     />
     <title>GOE's CINEMA</title>
     <link rel="stylesheet" type="text/css" href="../css/html.css" />
+    <script src="/topRated/topRated.js" type="module"></script>
   </head>
   <body>
     <header>
@@ -77,6 +78,5 @@
       </div>
     </footer>
     <div class="moveTopBtn"><img src="../img/top.png" alt="" /></div>
-    <script src="/topRated/topRated.js"></script>
   </body>
 </html>
