+++ upcoming/upcoming.html
@@ -12,6 +12,7 @@
     />
     <title>GOE's CINEMA</title>
     <link rel="stylesheet" type="text/css" href="../css/html.css" />
+    <script src="/upcoming/upcoming.js" type="module"></script>
   </head>
   <body>
     <header>
@@ -75,6 +76,5 @@
       </div>
     </footer>
     <div class="moveTopBtn"><img src="../img/top.png" alt="" /></div>
-    <script src="/upcoming/upcoming.js"></script>
   </body>
 </html>
